requirements.txt
asgiref==3.4.1
Django==3.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
Pillow==7.0.0
psycopg2-binary==2.8.4
PyJWT==2.2.0
pytz==2021.3
sqlparse==0.4.2
